Refatored 4 View Controllers (MovieListVC and four categyMoviesListVC) and created one ReusableMovieListVC.
CodnitionalFetchMethod check which data has been passed from SearchVC and call corresponding movie fetch methods (either searchMovie or fetchMovie).
SearchMovie method is for fetching data by a keyword. FetchMovie method is for fetching movies from a particular category.
