+++ PeliculaApp/Custom Views/Cells/CastCustomCell.swift
@@ -43,6 +43,7 @@ class CastCustomCell: UICollectionViewCell {
         castImageView.layer.cornerRadius = 60/2
         castImageView.clipsToBounds      = true
         castImageView.contentMode        = .scaleAspectFill
+        castImageView.image              = UIImage(named: "placeholder")
     }
